refactor: move `compat` into `lib`
It's also used by some files in `lib`, and we want `lib` to be free
of dependencies on other packages.
