+++ src/mb_enhanced_cover_art_uploads/fetch.ts
@@ -2,7 +2,7 @@ import { LOGGER } from '@lib/logging/logger';
 import { ArtworkTypeIDs } from '@lib/MB/CoverArt';
 import { gmxhr } from '@lib/util/xhr';
 import { urlBasename } from '@lib/util/urls';
-import { getFromPageContext } from '@src/compat';
+import { getFromPageContext } from '@lib/compat';
 import { getMaximisedCandidates } from './maximise';
 import { getProvider } from './providers';
 import type { CoverArt, CoverArtProvider } from './providers/base';
